部分完成了client_core 和 server_core
