+++ task1/model/coin.py
@@ -1,4 +1,6 @@
 from datetime import datetime
+
+import base64
 import uuid
 
 
@@ -8,6 +10,7 @@ class Coin:
     def __init__(self, expiry_date: datetime = None) -> None:
         self._uid = uuid.uuid4()
         self.expiry_date = expiry_date
+        self.signature: bytes = None
 
     @property
     def uid(self):
